x11-drivers/ati-drivers: Revision bump to 15.12-r3 with new patches
Add patches from Arch's AUR to support kernels through 4.15. Now:

-r1: Last official Gentoo version, supports kernels through 4.5
-r2: Patches for kernels between 4.6 and 4.9 LTS
-r3: Patches for kernels between 4.10 and 4.14 LTS, should also
     work with 4.15 according to AUR comments

Future: As long as patches are available, will probably move to one
revbump per LTS kernel. So -r3 will stay as is, and patches needed
to build on 4.16+ will be added to a new -r4, until the next LTS
kernel is released, at which point -r5 would be created.